Changed backends module organization. Renamed mpi test script to avoid autoexecution on make test
