+++ abcpy/backends/__init__.py
@@ -1,0 +1,3 @@
+from abcpy.backends.base import *
+from abcpy.backends.mpi import BackendMPI
+from  abcpy.backends.spark import BackendSpark
